[plugins] fastcgi: Record::size() shall return integer of 32 bit instead 16 to avoid overflow
The content lengths itself is already an uint16_t so the sum of this
plus padding plus header must be greater - an uint32_t.

+++ plugins/fastcgi_protocol.h
@@ -77,7 +77,7 @@ public:
 	const char *content() const { return reinterpret_cast<const char *>(this) + sizeof(*this); }
 
 	const char *data() const { return reinterpret_cast<const char *>(this); }
-	uint16_t size() const { return sizeof(Record) + contentLength() + paddingLength(); }
+	uint32_t size() const { return sizeof(Record) + contentLength() + paddingLength(); }
 
 	bool isManagement() const { return requestId() == 0; }
 	bool isApplication() const { return requestId() != 0; }
